+++ src/Symfony/Component/Notifier/CHANGELOG.md
@@ -5,11 +5,12 @@ CHANGELOG
 -----
 
  * [BC BREAK] The `TransportInterface::send()` and `AbstractTransport::doSend()` methods changed to return a `?SentMessage` instance instead of `void`.
+ * Added the Zulip notifier bridge
 
 5.1.0
 -----
 
- * Added the Mattermost notifier bridge
+* Added the Mattermost notifier bridge
 * [BC BREAK] The `ChatMessage::fromNotification()` method's `$recipient` and `$transport`
   arguments were removed.
 * [BC BREAK] The `EmailMessage::fromNotification()` and `SmsMessage::fromNotification()`